+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Set `isFollowed` in `getFollowedAccounts()`, always `Account.build()` starts `isFollowed` as false
 * Fixes to Toot.dedupeToots(), which was kind of broken because of uniquifying on retoot `realURI()` and randomly choosing retoots or real toots in final step
 * Lowercase `Account.buildWebfingerURI()`
-* Bump up minRecordsForFeatureScoring to 480
+* Bump up `minRecordsForFeatureScoring` to 480
+* Load Algorithm.homeFeed from cache on startup
 
 ### v0.30.1
 * Call `Toot.complete()` in batches of 50 instead of 100, increase sleep time to 250ms from 200